mac utility fix

+++ src/helpers/utility.ts
@@ -10,7 +10,13 @@ let utiltiy = {
 
     // empty out installer dir
     fse.readdir(installerDir, (err, files) => {
-      if (err) logger.error(err);
+      if (err){
+        logger.error(err);
+        return;
+      } 
+      
+      // if object is empty, exit mac fix
+      if(Object.keys(files).length === 0) return;
     
       for (let file of files) {
         fse.unlink(path.join(installerDir, file), err => {
@@ -21,12 +27,15 @@ let utiltiy = {
 
     // delete installers from app dir
     fse.readdir(appDir, (err, files) => {
-      if (err) logger.error(err);
+      if (err){
+        logger.error(err);
+        return;
+      }
     
       for (let file of files) {
         let regexp = /installer/gi;
         //let regexp = '/installer/gi';
-        if(file.match(regexp)){console.log(file);
+        if(file.match(regexp)){
           fse.unlink(path.join(appDir, file), err => {
             if (err) logger.error(err);
           });
